Refactored code to increase readability
