+++ src/const/ages.ts
@@ -1,0 +1,14 @@
+import { ValuedDescriptor } from '../models/valuedDescriptor'
+
+export const Ages: ValuedDescriptor[] = [
+    { weight: 1, label: 'Avant-garde', value: -3 },
+    { weight: 1, label: 'Newly-forged', value: -2 },
+    { weight: 4, label: 'Recent past', value: -1 },
+    { weight: 7, label: 'Old', value: 0 },
+    { weight: 3, label: 'Antique', value: 1 },
+    { weight: 2, label: 'Historical', value: 2 },
+    { weight: 1, label: 'Venerable', value: 3 },
+    { weight: 1, label: 'Archaic', value: 4 },
+    { weight: 1, label: 'Adetiluvian', value: 5 },
+    { weight: 1, label: 'Primordial', value: 6 },
+]
