+++ src/const/descriptors.ts
@@ -1,0 +1,238 @@
+export const Descriptors: string[] = [
+    'Engraved',
+    'Animated',
+    'Attractive',
+    'Adorned',
+    'Tactile',
+    'Austere',
+    'Authentic',
+    'Bare',
+    'Barren',
+    'Beautiful',
+    'Burdensome',
+    'Bold',
+    'Breakable',
+    'Cheap',
+    'Creamy',
+    'Colourful',
+    'Cold',
+    'Smooth',
+    'Cylindrical',
+    'Curly ',
+    'Damaged',
+    'Chipped',
+    'Damp',
+    'Dazzling',
+    'Crooked',
+    'Warm',
+    'Rough',
+    'Ragged',
+    'Slick',
+    'Dirty',
+    'Disfigured',
+    'Small',
+    'Distorted',
+    'Drab',
+    'Disguised',
+    'Dense',
+    'Delightful',
+    'Dimpled',
+    'Embellished',
+    'Even',
+    'Exotic',
+    'Expert',
+    'Expensive',
+    'Enchanted',
+    'Elliptical',
+    'Elastic',
+    'Edible',
+    'Exemplary',
+    'Evil',
+    'Pure',
+    'Feathered',
+    'Drow',
+    'Tortle',
+    'Elven',
+    'Dwarven',
+    'Abyssal',
+    'Demonic',
+    'Chantry',
+    'Religious',
+    'Fluid',
+    'Flaky',
+    'Functional',
+    'Funny',
+    'Gigantic',
+    'Glaring',
+    'Granular',
+    'Green',
+    'Yellow',
+    'Purple',
+    'Blue',
+    'Red',
+    'Orange',
+    'Pink',
+    'White',
+    'Black',
+    'Blank',
+    'Multicoloured',
+    'Grey',
+    'Brown',
+    'Aquamarine',
+    'Gruesoime',
+    'Gorgeous',
+    'Gentle',
+    'Grubby',
+    'Hairy',
+    'Hideous',
+    'Homely',
+    'Humble',
+    'Heavy',
+    'Light',
+    'Humming',
+    'Glowing',
+    'Imperfect',
+    'Impure',
+    'Ironclad',
+    'Inferior',
+    'Jaunty',
+    'Jagged',
+    'Juvenile',
+    'Kooky',
+    'Knotted',
+    'Little',
+    'Lavish',
+    'Luxurious',
+    'Low',
+    'Tall',
+    'Luminous',
+    'Majestic',
+    'Masculine',
+    'Feminine',
+    'Minty',
+    'Measly',
+    'Miniature',
+    'Modest',
+    'Mushy',
+    'Moist',
+    'Muddy',
+    'Musty',
+    'Mysterious',
+    'Neglected',
+    'Nautical',
+    'Novel',
+    'Natural',
+    'Obvious',
+    'Oblong',
+    'Odd',
+    'Offensive',
+    'Outlandish',
+    'Outrageous',
+    'Official',
+    'Old-fashioned',
+    'Optimal',
+    'Oddball',
+    'Peaceful',
+    'Pessimistic',
+    'Physical',
+    'Pitiful',
+    'Perky',
+    'Perfect',
+    'Pastel',
+    'Peppery',
+    'Piercing',
+    'Practical',
+    'Pristine',
+    'Powerful',
+    'Positive',
+    'Popular',
+    'Putrid',
+    'Private',
+    'Posh',
+    'Pleasant',
+    'Plain',
+    'Quaint',
+    'Quirky',
+    'Questionable',
+    'Radical',
+    'Regal',
+    'Revolting',
+    'Regular',
+    'Remarkable',
+    'Rigid',
+    'Rich',
+    'Rude',
+    'Robbed',
+    'Sad',
+    'Scrawny',
+    'Scholarly',
+    'Bookish',
+    'Sentient',
+    'Silky',
+    'Shameful',
+    'Shallow',
+    'Silver',
+    'Sentient',
+    'Slimy',
+    'Smooth',
+    'Smart',
+    'Shiny',
+    'Shocking',
+    'Serene',
+    'Sentimental',
+    'Second-hand',
+    'Slight',
+    'Squiggly',
+    'Squeaky',
+    'Sparse',
+    'Splendid',
+    'Sophisticated',
+    'Superb',
+    'Striking',
+    'Subdued',
+    'Talkative',
+    'Tempting',
+    'Terrific',
+    'Tuddy',
+    'Trimmed',
+    'Tinted',
+    'Threadbare',
+    'Ugly',
+    'Unfinished',
+    'Uniform',
+    'Uncommon',
+    'Unwieldy',
+    'Unrealistic',
+    'Unkempt',
+    'Useless',
+    'Usable',
+    'Untidy',
+    'Vast',
+    'Velvety',
+    'Vibrant',
+    'Wasteful',
+    'Wealthy',
+    'Warlike',
+    'Weird',
+    'Warm',
+    'Waterlogged',
+    'Whole',
+    'Worn',
+    'Wretched',
+    'Winged',
+    'Wilted',
+    'Yawning',
+    'Zany',
+    'Zesty',
+    'Dazzling',
+    'Precious',
+    'Barbed',
+    'Dark',
+    'Light',
+    'Double',
+    'Golden',
+    'Serrated',
+    'Vicious',
+    'Patterned',
+    'Banded',
+]
