+++ src/const/itemTypes.ts
@@ -1,0 +1,44 @@
+export const ItemTypes: string[] = [
+    'Art - caligraphy',
+    'Art - Manuscript',
+    'Art - Embroidery',
+    'Art - Cushion',
+    'Art - Wood relief carvings',
+    'Art - Printed textiles',
+    'Painting - Watercolours',
+    'Painting - Oils',
+    'Painting - Enamels',
+    'Painting - Watercolours',
+    'Painting - Oils',
+    'Painting - Enamels',
+    'Ceramings - Pottery',
+    'Ceramics - Urns',
+    'Ceramics - chinaware',
+    'Ceramics - figurine',
+    'Ceramics - figurine',
+    'Glass - Decanters',
+    'Glass - goblets',
+    'Glass - bowls',
+    'Glass - engravings',
+    'Stone - Statuette',
+    'Stone - Bust',
+    'Stone - Plaque',
+    'Stone - Carving',
+    'Metal - Sculpture',
+    'Metal - Furnishing',
+    'Metal - Cutlary',
+    'Metal - Tools',
+    'Jewellery - Ring',
+    'Jewellery - Necklace',
+    'Jewellery - Brooch',
+    'Jewellery - Clasp',
+    'Jewellery - Ring',
+    'Jewellery - Necklace',
+    'Jewellery - Brooch',
+    'Jewellery - Pin',
+    'Jewellery - Wooden Box',
+    'Jewellery - Pendant',
+    'Jewellery - Tiara/Crown',
+    'MAGIC ITEM',
+    'MAGIC ITEM',
+]
