+++ src/const/materials.ts
@@ -1,0 +1,34 @@
+export const Materials: string[] = [
+    'Wood - Light Birch',
+    'Wood - Pine',
+    'Wood - Cypress',
+    'Wood - Cedar',
+    'Wood - Redcedar',
+    'Wood - Solid Oak',
+    'Wood - Grey Duskwood',
+    'Wood - Silverbark',
+    'Fabric - Wool',
+    'Fabric - Cashmere',
+    'Fabric - Cotton',
+    'Fabric - Hemp/Woven',
+    'Fabric - Silk',
+    'Fabric - Burlap',
+    'Fabric - Fur',
+    'Fabric - Satin',
+    'Metal - Bronze',
+    'Metal - Platinum',
+    'Metal - Silver',
+    'Metal - Gold',
+    'Metal - Nickel',
+    'Metal - Adamantione',
+    'Metal - Green Steel',
+    'Metal - Steel',
+    'Metal - Lead',
+    'Metal - Coral',
+    'Stone - Marble',
+    'Stone - Basalt',
+    'Stone - Obsidian',
+    'Stone - Quartz',
+    'Stone - Sandstone',
+    'Stone - Limestone',
+]
