+++ src/const/sizes.ts
@@ -1,0 +1,13 @@
+import { ValuedDescriptor } from '../models/valuedDescriptor'
+
+export const Sizes: ValuedDescriptor[] = [
+    { weight: 2, label: 'Tiny', value: -3 },
+    { weight: 2, label: 'Small', value: -2 },
+    { weight: 2, label: 'Below Avg.', value: -1 },
+    { weight: 7, label: 'Average', value: 0 },
+    { weight: 3, label: 'Above Avg.', value: 1 },
+    { weight: 2, label: 'Large', value: 2 },
+    { weight: 1, label: 'Very Large', value: 3 },
+    { weight: 1, label: 'Huge', value: 4 },
+    { weight: 1, label: 'Gargantuan', value: 5 },
+]
