+++ src/models/item.ts
@@ -1,3 +1,25 @@
+import { getRandomValueFromArray, createdWeightedList } from '../utils'
+import { ValuedDescriptor } from './valuedDescriptor'
+import { Materials } from '../const/materials'
+import { Colours } from '../const/colours'
+import { Descriptors } from '../const/descriptors'
+import { Conditions } from '../const/conditions'
+import { Sizes } from '../const/sizes'
+import { Renowns } from '../const/renowns'
+import { Ages } from '../const/ages'
+import { ItemTypes } from '../const/itemTypes'
+
+export interface IItem {
+    type: string
+    material: string
+    colour: string
+    descriptors: string[]
+    condition: ValuedDescriptor
+    size: ValuedDescriptor
+    renown: ValuedDescriptor
+    age: ValuedDescriptor
+}
+
 export class Item {
     type: string
     material: string
@@ -8,24 +30,35 @@ export class Item {
     renown: ValuedDescriptor
     age: ValuedDescriptor
 
-    constructor(
-        type: string,
-        material: string,
-        colour: string,
-        descriptors: string[],
-        condition: ValuedDescriptor,
-        size: ValuedDescriptor,
-        renown: ValuedDescriptor,
-        age: ValuedDescriptor
-    ) {
-        this.type = type
-        this.material = material
-        this.colour = colour
-        this.descriptors = descriptors
-        this.condition = condition
-        this.size = size
-        this.renown = renown
-        this.age = age
+    constructor()
+    constructor(item: IItem)
+    constructor(item?: IItem) {
+        this.type = item?.type ? item.type : getRandomValueFromArray(ItemTypes)
+        this.material = item?.material
+            ? item.material
+            : getRandomValueFromArray(Materials)
+        this.colour = item?.colour
+            ? item.colour
+            : getRandomValueFromArray(Colours)
+        this.descriptors = item?.descriptors
+            ? item.descriptors
+            : [
+                  getRandomValueFromArray(Descriptors),
+                  getRandomValueFromArray(Descriptors),
+                  getRandomValueFromArray(Descriptors),
+              ]
+        this.condition = item?.condition
+            ? item.condition
+            : getRandomValueFromArray(createdWeightedList(Conditions))
+        this.size = item?.size
+            ? item.size
+            : getRandomValueFromArray(createdWeightedList(Sizes))
+        this.renown = item?.renown
+            ? item.renown
+            : getRandomValueFromArray(createdWeightedList(Renowns))
+        this.age = item?.age
+            ? item.age
+            : getRandomValueFromArray(createdWeightedList(Ages))
     }
 
     formatItemForMessage(): string {
@@ -59,519 +92,3 @@ export class Item {
         )
     }
 }
-
-export type ValuedDescriptor = {
-    value: number
-    label: string
-    weight: number
-}
-
-export const Descriptors: string[] = [
-    'Engraved',
-    'Animated',
-    'Attractive',
-    'Adorned',
-    'Tactile',
-    'Austere',
-    'Authentic',
-    'Bare',
-    'Barren',
-    'Beautiful',
-    'Burdensome',
-    'Bold',
-    'Breakable',
-    'Cheap',
-    'Creamy',
-    'Colourful',
-    'Cold',
-    'Smooth',
-    'Cylindrical',
-    'Curly ',
-    'Damaged',
-    'Chipped',
-    'Damp',
-    'Dazzling',
-    'Crooked',
-    'Warm',
-    'Rough',
-    'Ragged',
-    'Slick',
-    'Dirty',
-    'Disfigured',
-    'Small',
-    'Distorted',
-    'Drab',
-    'Disguised',
-    'Dense',
-    'Delightful',
-    'Dimpled',
-    'Embellished',
-    'Even',
-    'Exotic',
-    'Expert',
-    'Expensive',
-    'Enchanted',
-    'Elliptical',
-    'Elastic',
-    'Edible',
-    'Exemplary',
-    'Evil',
-    'Pure',
-    'Feathered',
-    'Drow',
-    'Tortle',
-    'Elven',
-    'Dwarven',
-    'Abyssal',
-    'Demonic',
-    'Chantry',
-    'Religious',
-    'Fluid',
-    'Flaky',
-    'Functional',
-    'Funny',
-    'Gigantic',
-    'Glaring',
-    'Granular',
-    'Green',
-    'Yellow',
-    'Purple',
-    'Blue',
-    'Red',
-    'Orange',
-    'Pink',
-    'White',
-    'Black',
-    'Blank',
-    'Multicoloured',
-    'Grey',
-    'Brown',
-    'Aquamarine',
-    'Gruesoime',
-    'Gorgeous',
-    'Gentle',
-    'Grubby',
-    'Hairy',
-    'Hideous',
-    'Homely',
-    'Humble',
-    'Heavy',
-    'Light',
-    'Humming',
-    'Glowing',
-    'Imperfect',
-    'Impure',
-    'Ironclad',
-    'Inferior',
-    'Jaunty',
-    'Jagged',
-    'Juvenile',
-    'Kooky',
-    'Knotted',
-    'Little',
-    'Lavish',
-    'Luxurious',
-    'Low',
-    'Tall',
-    'Luminous',
-    'Majestic',
-    'Masculine',
-    'Feminine',
-    'Minty',
-    'Measly',
-    'Miniature',
-    'Modest',
-    'Mushy',
-    'Moist',
-    'Muddy',
-    'Musty',
-    'Mysterious',
-    'Neglected',
-    'Nautical',
-    'Novel',
-    'Natural',
-    'Obvious',
-    'Oblong',
-    'Odd',
-    'Offensive',
-    'Outlandish',
-    'Outrageous',
-    'Official',
-    'Old-fashioned',
-    'Optimal',
-    'Oddball',
-    'Peaceful',
-    'Pessimistic',
-    'Physical',
-    'Pitiful',
-    'Perky',
-    'Perfect',
-    'Pastel',
-    'Peppery',
-    'Piercing',
-    'Practical',
-    'Pristine',
-    'Powerful',
-    'Positive',
-    'Popular',
-    'Putrid',
-    'Private',
-    'Posh',
-    'Pleasant',
-    'Plain',
-    'Quaint',
-    'Quirky',
-    'Questionable',
-    'Radical',
-    'Regal',
-    'Revolting',
-    'Regular',
-    'Remarkable',
-    'Rigid',
-    'Rich',
-    'Rude',
-    'Robbed',
-    'Sad',
-    'Scrawny',
-    'Scholarly',
-    'Bookish',
-    'Sentient',
-    'Silky',
-    'Shameful',
-    'Shallow',
-    'Silver',
-    'Sentient',
-    'Slimy',
-    'Smooth',
-    'Smart',
-    'Shiny',
-    'Shocking',
-    'Serene',
-    'Sentimental',
-    'Second-hand',
-    'Slight',
-    'Squiggly',
-    'Squeaky',
-    'Sparse',
-    'Splendid',
-    'Sophisticated',
-    'Superb',
-    'Striking',
-    'Subdued',
-    'Talkative',
-    'Tempting',
-    'Terrific',
-    'Tuddy',
-    'Trimmed',
-    'Tinted',
-    'Threadbare',
-    'Ugly',
-    'Unfinished',
-    'Uniform',
-    'Uncommon',
-    'Unwieldy',
-    'Unrealistic',
-    'Unkempt',
-    'Useless',
-    'Usable',
-    'Untidy',
-    'Vast',
-    'Velvety',
-    'Vibrant',
-    'Wasteful',
-    'Wealthy',
-    'Warlike',
-    'Weird',
-    'Warm',
-    'Waterlogged',
-    'Whole',
-    'Worn',
-    'Wretched',
-    'Winged',
-    'Wilted',
-    'Yawning',
-    'Zany',
-    'Zesty',
-    'Dazzling',
-    'Precious',
-    'Barbed',
-    'Dark',
-    'Light',
-    'Double',
-    'Golden',
-    'Serrated',
-    'Vicious',
-    'Patterned',
-    'Banded',
-]
-
-export const Types: string[] = [
-    'Art - caligraphy',
-    'Art - Manuscript',
-    'Art - Embroidery',
-    'Art - Cushion',
-    'Art - Wood relief carvings',
-    'Art - Printed textiles',
-    'Painting - Watercolours',
-    'Painting - Oils',
-    'Painting - Enamels',
-    'Painting - Watercolours',
-    'Painting - Oils',
-    'Painting - Enamels',
-    'Ceramings - Pottery',
-    'Ceramics - Urns',
-    'Ceramics - chinaware',
-    'Ceramics - figurine',
-    'Ceramics - figurine',
-    'Glass - Decanters',
-    'Glass - goblets',
-    'Glass - bowls',
-    'Glass - engravings',
-    'Stone - Statuette',
-    'Stone - Bust',
-    'Stone - Plaque',
-    'Stone - Carving',
-    'Metal - Sculpture',
-    'Metal - Furnishing',
-    'Metal - Cutlary',
-    'Metal - Tools',
-    'Jewellery - Ring',
-    'Jewellery - Necklace',
-    'Jewellery - Brooch',
-    'Jewellery - Clasp',
-    'Jewellery - Ring',
-    'Jewellery - Necklace',
-    'Jewellery - Brooch',
-    'Jewellery - Pin',
-    'Jewellery - Wooden Box',
-    'Jewellery - Pendant',
-    'Jewellery - Tiara/Crown',
-    'MAGIC ITEM',
-    'MAGIC ITEM',
-]
-
-export const Materials: string[] = [
-    'Wood - Light Birch',
-    'Wood - Pine',
-    'Wood - Cypress',
-    'Wood - Cedar',
-    'Wood - Redcedar',
-    'Wood - Solid Oak',
-    'Wood - Grey Duskwood',
-    'Wood - Silverbark',
-    'Fabric - Wool',
-    'Fabric - Cashmere',
-    'Fabric - Cotton',
-    'Fabric - Hemp/Woven',
-    'Fabric - Silk',
-    'Fabric - Burlap',
-    'Fabric - Fur',
-    'Fabric - Satin',
-    'Metal - Bronze',
-    'Metal - Platinum',
-    'Metal - Silver',
-    'Metal - Gold',
-    'Metal - Nickel',
-    'Metal - Adamantione',
-    'Metal - Green Steel',
-    'Metal - Steel',
-    'Metal - Lead',
-    'Metal - Coral',
-    'Stone - Marble',
-    'Stone - Basalt',
-    'Stone - Obsidian',
-    'Stone - Quartz',
-    'Stone - Sandstone',
-    'Stone - Limestone',
-]
-export const Colours: string[] = [
-    'MediumVioletRed',
-    'DeepPink',
-    'PaleVioletRed',
-    'HotPink',
-    'LightPink',
-    'Pink',
-    'DarkRed',
-    'Red',
-    'Firebrick',
-    'Crimson',
-    'IndianRed',
-    'LightCoral',
-    'Salmon',
-    'DarkSalmon',
-    'LightSalmon',
-    'OrangeRed',
-    'Tomato',
-    'DarkOrange',
-    'Coral',
-    'Orange',
-    'DarkKhaki',
-    'Gold',
-    'Khaki',
-    'PeachPuff',
-    'Yellow',
-    'PaleGoldenrod',
-    'Moccasin',
-    'PapayaWhip',
-    'LightGoldenrodYellow',
-    'LemonChiffon',
-    'LightYellow',
-    'Maroon',
-    'Brown',
-    'SaddleBrown',
-    'Sienna',
-    'Chocolate',
-    'DarkGoldenrod',
-    'Peru',
-    'RosyBrown',
-    'Goldenrod',
-    'SandyBrown',
-    'Tan',
-    'Burlywood',
-    'Wheat',
-    'NavajoWhite',
-    'Bisque',
-    'BlanchedAlmond',
-    'Cornsilk',
-    'DarkGreen',
-    'Green',
-    'DarkOliveGreen',
-    'ForestGreen',
-    'SeaGreen',
-    'Olive',
-    'OliveDrab',
-    'MediumSeaGreen',
-    'LimeGreen',
-    'Lime',
-    'SpringGreen',
-    'MediumSpringGreen',
-    'DarkSeaGreen',
-    'MediumAquamarine',
-    'YellowGreen',
-    'LawnGreen',
-    'Chartreuse',
-    'LightGreen',
-    'GreenYellow',
-    'PaleGreen',
-    'Teal',
-    'DarkCyan',
-    'LightSeaGreen',
-    'CadetBlue',
-    'DarkTurquoise',
-    'MediumTurquoise',
-    'Turquoise',
-    'Aqua',
-    'Cyan',
-    'Aquamarine',
-    'PaleTurquoise',
-    'LightCyan',
-    'Navy',
-    'DarkBlue',
-    'MediumBlue',
-    'Blue',
-    'MidnightBlue',
-    'RoyalBlue',
-    'SteelBlue',
-    'DodgerBlue',
-    'DeepSkyBlue',
-    'CornflowerBlue',
-    'SkyBlue',
-    'LightSkyBlue',
-    'LightSteelBlue',
-    'LightBlue',
-    'PowderBlue',
-    'Indigo',
-    'Purple',
-    'DarkMagenta',
-    'DarkViolet',
-    'DarkSlateBlue',
-    'BlueViolet',
-    'DarkOrchid',
-    'Fuchsia',
-    'Magenta',
-    'SlateBlue',
-    'MediumSlateBlue',
-    'MediumOrchid',
-    'MediumPurple',
-    'Orchid',
-    'Violet',
-    'Plum',
-    'Thistle',
-    'Lavender',
-    'MistyRose',
-    'AntiqueWhite',
-    'Linen',
-    'Beige',
-    'WhiteSmoke',
-    'LavenderBlush',
-    'OldLace',
-    'AliceBlue',
-    'Seashell',
-    'GhostWhite',
-    'Honeydew',
-    'FloralWhite',
-    'Azure',
-    'MintCream',
-    'Snow',
-    'Ivory',
-    'White',
-    'Black',
-    'DarkSlateGray',
-    'DimGray',
-    'SlateGray',
-    'Gray',
-    'LightSlateGray',
-    'DarkGray',
-    'Silver',
-    'LightGray',
-    'Gainsboro',
-]
-
-export const Conditions: ValuedDescriptor[] = [
-    { weight: 1, label: 'Awful', value: -4 },
-    { weight: 1, label: 'Very Poor', value: -3 },
-    { weight: 2, label: 'Poor', value: -2 },
-    { weight: 2, label: 'Below Avg.', value: -1 },
-    { weight: 7, label: 'Average', value: 0 },
-    { weight: 4, label: 'Above Avg.', value: 1 },
-    { weight: 1, label: 'Good', value: 2 },
-    { weight: 1, label: 'Very Good', value: 3 },
-    { weight: 1, label: 'Excellent', value: 4 },
-    { weight: 1, label: 'Flawless', value: 5 },
-]
-export const Sizes: ValuedDescriptor[] = [
-    { weight: 2, label: 'Tiny', value: -3 },
-    { weight: 2, label: 'Small', value: -2 },
-    { weight: 2, label: 'Below Avg.', value: -1 },
-    { weight: 7, label: 'Average', value: 0 },
-    { weight: 3, label: 'Above Avg.', value: 1 },
-    { weight: 2, label: 'Large', value: 2 },
-    { weight: 1, label: 'Very Large', value: 3 },
-    { weight: 1, label: 'Huge', value: 4 },
-    { weight: 1, label: 'Gargantuan', value: 5 },
-]
-
-export const Renowns: ValuedDescriptor[] = [
-    { weight: 3, label: 'Unknown', value: -3 },
-    { weight: 3, label: 'Obscurely-known', value: -2 },
-    { weight: 1, label: 'Obscurely-known', value: -1 },
-    { weight: 6, label: 'Regionally-known', value: 0 },
-    { weight: 4, label: 'Nationally-known', value: 1 },
-    { weight: 1, label: 'Continentally-known', value: 2 },
-    { weight: 1, label: 'World-known', value: 3 },
-    { weight: 1, label: 'Fabled', value: 4 },
-    { weight: 1, label: 'Legendary', value: 5 },
-]
-
-export const Ages: ValuedDescriptor[] = [
-    { weight: 1, label: 'Avant-garde', value: -3 },
-    { weight: 1, label: 'Newly-forged', value: -2 },
-    { weight: 4, label: 'Recent past', value: -1 },
-    { weight: 7, label: 'Old', value: 0 },
-    { weight: 3, label: 'Antique', value: 1 },
-    { weight: 2, label: 'Historical', value: 2 },
-    { weight: 1, label: 'Venerable', value: 3 },
-    { weight: 1, label: 'Archaic', value: 4 },
-    { weight: 1, label: 'Adetiluvian', value: 5 },
-    { weight: 1, label: 'Primordial', value: 6 },
-]
