+++ src/models/valuedDescriptor.ts
@@ -1,0 +1,5 @@
+export type ValuedDescriptor = {
+    value: number
+    label: string
+    weight: number
+}
