+++ src/utils.ts
@@ -1,22 +1,21 @@
 import { Message } from 'discord.js'
-import { attack } from './commands/attack'
-import { ping } from './commands/ping'
-import config from './config.json'
-import { snack } from './commands/snack'
-import { defend } from './commands/defend'
-import { speak } from './commands/speak'
-import _do from './commands/do'
-import { logger } from './logger'
-import { setConfig } from './commands/setConfig'
-import { getConfig } from './commands/getConfig'
-import { GuildSettings, IGuild } from './models/guild'
-import Guild, { IGuildModel } from './models/guild'
-import die from './commands/dies'
-import en from 'javascript-time-ago/locale/en'
 import TimeAgo from 'javascript-time-ago'
-import resurrect from './commands/resurrect'
-import { ValuedDescriptor } from './models/item'
+import en from 'javascript-time-ago/locale/en'
+import { attack } from './commands/attack'
+import { defend } from './commands/defend'
+import die from './commands/dies'
+import _do from './commands/do'
 import { generateItem } from './commands/generateItem'
+import { getConfig } from './commands/getConfig'
+import { ping } from './commands/ping'
+import resurrect from './commands/resurrect'
+import { setConfig } from './commands/setConfig'
+import { snack } from './commands/snack'
+import { speak } from './commands/speak'
+import config from './config.json'
+import { logger } from './logger'
+import Guild, { GuildSettings } from './models/guild'
+import { ValuedDescriptor } from './models/valuedDescriptor'
 
 // set up TimeAgo
 TimeAgo.addLocale(en)
